requirements.txt
numpy
pandas
padelpy
rdkit-pypi
chembl_webresource_client
sklearn
matplotlib
seaborn